refactor: enum값들도 각 패키지에 분류
