+++ src/main/java/umc/lightup/member/domain/Member.java
@@ -2,8 +2,8 @@ package umc.lightup.member.domain;
 
 import jakarta.persistence.*;
 import umc.lightup.common.BaseEntity;
-import umc.lightup.common.enums.Mbti;
-import umc.lightup.common.enums.Role;
+import umc.lightup.member.enums.Mbti;
+import umc.lightup.member.enums.Role;
 
 import java.time.LocalDate;
