+++ src/main/java/umc/lightup/notification/enums/ReferenceType.java
@@ -1,4 +1,4 @@
-package umc.lightup.common.enums;
+package umc.lightup.notification.enums;
 
 public enum ReferenceType {
     ITEM, MEMBER, PROJECT // 실제 타입에 맞게 추가
