+++ src/main/java/umc/lightup/tool/enums/ToolType.java
@@ -1,4 +1,4 @@
-package umc.lightup.common.enums;
+package umc.lightup.tool.enums;
 
 public enum ToolType {
     SOFTWARE, HARDWARE // 실제 타입에 맞게 추가
